update. screens navigation
